Fix warning (it was supposed to be 32-bit, not 64)

+++ TDR/fixedpoint.cpp
@@ -23,10 +23,10 @@ fix1616 tdr::fix1616::operator-(fix1616 b) const {
 	return { u - b.u };
 }
 fix1616 tdr::fix1616::addWrap(fix1616 b) const {
-  return { (int64_t) ((uint64_t) u + (uint64_t) b.u) };
+  return { (int32_t) ((uint32_t) u + (uint32_t) b.u) };
 }
 fix1616 tdr::fix1616::subWrap(fix1616 b) const {
-  return { (int64_t) ((uint64_t) u - (uint64_t) b.u) };
+  return { (int32_t) ((uint32_t) u - (uint32_t) b.u) };
 }
 fix1616 tdr::fix1616::operator-() const {
 	if (u == ABYSS) return { CELESTIAL };
